MPfm 0.5.0.5 - Migrated Button to new CustomFont class.
Related to issue #221.


git-svn-id: http://hamster-svn/svn/repos/base@217 765c1f7c-9fb8-954f-9ff8-dd0915cb3117

+++ MPfm/branches/current/MPfm.WindowsControls/Fonts/CustomFontEditorForm.cs
@@ -117,7 +117,7 @@ namespace MPfm.WindowsControls
             // Set initial values
             radioUseCustomFont.Checked = CustomFont.UseEmbeddedFont;
             lblFontSize.Text = "Font Size: " + CustomFont.Size.ToString() + " pt";
-            trackFontSize.Value = CustomFont.Size;
+            trackFontSize.Value = (int)CustomFont.Size;
             chkIsBold.Checked = CustomFont.IsBold;
             chkIsItalic.Checked = CustomFont.IsItalic;
             chkIsUnderline.Checked = CustomFont.IsUnderline;
